Update translations from AppVeyor
Along with this change, the RC file in each translation project has been
split into three files:

- One that contains most of the original resource data (dialogs, menus
  and the string table)
- One that contains the version info data
- A resource file that simply includes the previous two files (it's this
  resource file that's actually built during compilation)

This matches the structure of the English resource file in the main
project and is needed to be able to easily update the translation
resources via AppVeyor. It also means that the main resource data can
change, while the version info data remains the same.

Note that along with the above changes, the main resource files now
include resource.h from the main project. This means that if a resource
constant is renamed or removed, the release build will fail (as the
translation projects will fail to compile). This shouldn't be too much
of an issue, as the updated translation files will just need to be
pulled in from AppVeyor.

There is also the case where items are added to the resource file in the
main project. That won't cause a build failure, but it will mean that
the translations will be out of date (and technically incompatible)
until they're updated.

With the way the translations are currently set up, it's not really
possible to avoid that and there's no guarantee made that the
translations will be compatible with each commit of a development build.
As in the above situation, the translation resources would need to be
updated from AppVeyor to remain compatible.