Temporarily disable code that handles file association changes
This is causing various shell icons in the application to not display
correctly (see issue #169 for some more information). It's especially
noticeable, as icons in the drives toolbar/address bar can disappear on
application startup.

As a temporary measure, the code that handles file association changes
(and attempts to update the icons the application uses) has been
disabled.