Only load language after loading settings
Previously (see 47bad0b51171f478408da30201fb5c07b4a231f3), the code that
loads the language a user selected was moved to an earlier point in the
application initialization. However, because of this, the language would
be initialized before the appropriate language settings had been loaded.
This meant that if you selected a language from the options dialog, when
you restarted, the language would still be English.

To fix this, the language initialization is now done after the settings
are loaded. This effectively reverts the commit referenced above.

Note that, with the way the color rules are currently set up, it doesn't
make much sense to load the default color rule strings from a language
DLL anyway.

The default color rules are specifically initialized before any settings
are loaded, which means that they can't then depend on settings that are
loaded. This could be changed so that the default color rules are only
set up if there are no color rules present in the saved settings. But
that should never be the case; if there are settings, there should be a
list of color rules (even if the list is empty).

Finally, when a color rule is saved, it's the description that's saved
(and not the original string table ID). So once the default color rules
are saved, the only thing present in the settings would be the original,
English, strings. Assuming a user will run the application at least once
to choose a language, that means the default color rules would only be
set up on the first run and then the rules would be loaded from the
settings after that.

To properly resolve this, the application might need to store some state
that indicates that the default rules haven't been changed. It still
might be surprising if the default rules are suddenly translated though,
as rules the user enters would never be translated. Given that, it's not
clear whether it's even worth doing at all.