Build main menu accelerator strings dynamically when plugin support is enabled
Using a plugin, it's possible to update the keyboard accelerators that
the application uses. These accelerators are also statically included in
the text shown for a menu item.

With this change, the accelerator string is now dynamically generated
for each of the main menu items when plugin support is enabled. This
means that if a plugin has changed an accelerator for one of those
items, the updated accelerator is what will be displayed alongside the
menu text.

Note the following limitations:

- Some of the main menu items have shortcut keys, but those keys aren't
  handled through the accelerator system, so won't be shown. For
  example, File -> Delete is associated with the delete key, but that
  key isn't handled as an accelerator, so it won't be shown next to the
  menu item text.
- Some menu items have multiple accelerators, so the one that's shown
  won't necessarily be the one that's normally shown. For example,
  File -> Close Tab is associated with the accelerators "Ctrl + W" and
  "Ctrl + F4". The first is what's normally displayed, while the second
  might be displayed when the accelerator strings are dynamically
  generated.
- If a plugin adds a menu item with the accelerator key combination
  embedded in the text, that key combination will be removed.