Set minimum version to Windows 7
This has some of the same benefits as removing support for Windows XP
(see 502d7a28cb287c0e28cff845c312d8fd00261b96). In particular, it means
that it's easier to use functionality introduced in Windows 7.

Additionally, about 0.1% of visitors to https://explorerplusplus.com/
over the last year (October 2018 - October 2019) were using Windows
Vista and the global market share of Vista (as a proportion of Windows'
total market share) was about 0.3% in September 2019. Which means this
change shouldn't have too much of an impact.