Update all projects to use Windows 10 SDK
The version of the SDK has been set to "10.0 (latest installed
version)". It's assumed that releases of the Windows 10 SDK will remain
backwards compatible. As no features introduced more recently than
Windows 7 will be used (at least not directly from the headers), it
shouldn't make a difference which version of the 10.0 SDK is actually
used.

Not using a specific version also helps to avoid issues with AppVeyor
(if Explorer++ relies on an SDK version that AppVeyor stops
pre-installing). It also means that other potential developers don't
need to have a specific version of the 10.0 SDK installed.

Because no specific version is required, it does present the potential
risk that there will be issues compiling Explorer++ with an older
version of the SDK if newer functionality (e.g. something like a new
macro) is used, either directly (by the application itself) or
indirectly (by one of the dependent libraries).