Remove VS2015 platform toolset declarations
The projects in the solution may still compile with those versions of
Visual Studio, but it's not something I've checked any time recently and
it's not something I intend to maintain.

Additionally, supporting older versions of Visual Studio out of the box
means including NuGet packages for Boost that target those versions of
Visual Studio. That means that developers that build the solution have
to download more packages, regardless of whether or not that those
packages actually relevant.

Supporting the two most recent versions of Visual Studio (or perhaps
only the most recent, depending on  the features required) is perhaps
the best solution and is what will be maintained going forward.