Mark application as per-monitor DPI aware (v2)
Note that not everything in the application currently scales with DPI
(the icons are a notable case), but marking the application as DPI aware
is likely better than not doing it. Without DPI awareness, Windows will
simply bitmap scale the application, which makes it look noticeably
blurry.

Additionally, various parts of the UI scale automatically (e.g. the
dialogs), so the application mostly looks ok under high DPI with DPI
awareness turned on.