Load bookmark icons using the system small icon size
Rather than hardcoding icon sizes, it makes sense to use the system icon
size when possible. This should result in the same outcome, but the
intention of the code should be clearer. Therefore, the bookmark icons
are now loaded using the system small icon size, rather than a hardcoded
size.

Note that with this change, the icon resource loader functions have also
been updated so that the accept a desired icon width and height.
Previously, the simply accepted a single size value.

Although all icon resources in the application are square and the
desired icons widths/heights should be the same, making this change
means that that assumption doesn't have to be true. The desired
width/height could vary and the scaling would still work correctly
(though the icons would be stretched).