Explicitly set tab index values during startup
During startup, tabs are restored in their original order. If no index
is explicitly set when calling TabContainer::CreateNewTab(), the index
of the new tab will be computed.

When the "Open new tabs next to the current one" option is set,
CreateNewTab() will attempt to retrieve the index of the current tab.
However, when the first tab is being restored, there is no current tab.
It doesn't make much sense to check this option during startup anyway,
since tabs are created in the necessary order anyway.

Therefore, each tab index is now set explicitly during startup. This
fixes #176.