Remove ApplyLoadedSettings()
The only thing this did was enable/disable hidden files in the treeview.
However, this setting would only be set on startup (and not at any other
time), even if you changed the default show hidden setting.

It would be better to rethink how to enable/disable the display of
hidden files in the treeview.