Fix LPITEMIDLIST __unaligned warnings
Though these warnings are unlikely to be a problem in practice (since
__unaligned is only relevant on Itanium; see
https://stackoverflow.com/q/50975437 for some discussion), the code has
been updated so that the __unaligned specifier is (indirectly) used when
declaring a smart pointer to an ITEMIDLIST type.

This is done by basing the type of the smart pointer on the type of the
original pointer. In this way, __unaligned will be implicitly included
when needed.