Don't update treeview selection when treeview is hidden
The process of updating the selection can be relatively expensive (it
can take a few hundred ms). This is because each ancestor folder along
the current path needs to be expanded and enumerated. Not doing this
when the treeview is hidden avoids unnecessary work.

If the treeview is shown again, the selection will be updated, so the
work is effectively deferred until it's needed.