Only update treeview selection once during startup
As mentioned in the previous commit, updating the treeview selection can
be relatively expensive and doing it multiple times on startup (once for
each tab) even more so. Therefore, doing it only once (for the current
tab) on startup should be beneficial.

For example, with 4 tabs open, all under the same parent folder,
updating the treeview selection once during startup (as opposed to
multiple times) saved about 50ms in release mode. The amount saved in
other circumstances will vary based on things like the number of tabs
that are being restored and their locations.