Update way in which CShellBrowser manages some directory-specific values
The CShellBrowser class is tied to a tab and an instance exists for as
long as the associated tab exists. However, there's data stored by the
class that needs to be reset whenever the directory changes.

Previously, all of this data has been stored in the class directly,
alongside all the other data that exists for the lifetime of the class.
This makes it hard to see exactly what data needs to be reset whenever
the directory changes and also means that individual items need to be
reset or cleared in a manual way.

To help alleviate this issue, some directory-specific values are now
stored in a dedicated struct and an instance of that struct is reset
when the directory changes.