Use ResourceHelper::LoadString() in place of LoadString()
The main advantage of using the ResourceHelper version is that the
string size can be allocated exactly. When calling LoadString()
directly, a fixed size buffer tends to be passed in. This means that a
reasonable maximum size has to be chosen upfront. This size might be
insufficient if the string is updated in the future or if a translated
version is significantly longer than the original English text.

Allocating the buffer based on the size of the string in the resource
table should therefore work better.

Two other minor advantages of using the ResourceHelper version are that
it takes fewer arguments and returns the string directly (rather than
through an out parameter).