Allow most recently closed tab to be restored
Note that, currently, while the tab state is restored, the state of the
CShellBrowser object isn't. That means that if the tab has a custom
name, for example, that name will be restored and the directory will be
correctly set. However, the other properties associated with the display
of the folder's items will all have default values. The back/forward
history for the tab will also be empty.

Tab restore has been assigned the standard shortcut of Ctrl+Shift+T.