Replace PathManager class with NavigationController class
The newer class is managed by the Tab class, rather than CShellBrowser.
This removes some redundancies in the CShellBrowser class and is
generally simpler. This approach should also work out better, because
CShellBrowser doesn't need to be able to navigate back/forward on its
own and so doesn't need to manage the class that stores history entries.