Move middle click handling code to CShellBrowser class
In addition to being moved, the code has also been simplified, so that
it's shorter and easier to read. In addition, it uses the cached item
attributes (to check whether or not the item that was middle clicked is
a directory), rather than checking the attributes set on the file.