Move navigationCompletedSignal to TabContainer class
The Navigation class is effectively redundant (NavigationController now
covers most of its functionality), so this commit forms part of the work
to remove the class altogether.

This change also simplifies how the navigation completed event is fired.
The CShellBrowser class will fire an event every time it completes a
navigation and the TabContainer class will observe this event and fire a
second event, allowing other classes to observe all navigations across
tabs.