Move ownership of NavigationController to CShellBrowser
This is ultimately necessary in order to allow a tab to navigate itself.
Although CShellBrowser::BrowseFolder() exists, it's not sufficient, for
a few reasons:

- The navigation mode (i.e. to determine whether or not a new tab should
  be created) is only checked by NavigationController.
- The NavigationController holds the back/forward history, so is needed
  to be able to navigate through the history.
- The method allowing the user to navigate up a directory is within
  NavigationController.

One advantage of this change was that the BrowseFolder() method was made
private within the CShellBrowser class. This means that the only way to
navigate a tab now is through the NavigationController. Previously, it
was possible to use the NavigationController or the BrowseFolder()
method directly (which some functions were doing).

Note that BrowseFolder() is still available internally within the
CShellBrowser class, even though it shouldn't be used directly. In the
future, it might be worth moving this method out of the class, so that
the NavigationController represents the only way to navigate at all.