Simplify handling of LVN_ITEMCHANGED display updates
Previously, the Explorerplusplus class kept track of whether the
listview selection was being changed in bulk. If this was the case, the
window states wouldn't be updated with every change notification;
instead they would be updated once the bulk update had finished.

However, this system meant that a certain amount of state had to be
tracked (increasing complexity and the chance of bugs) and only specific
bulk update mechanisms were handled. This meant that if the user
selected a group of items in another way (e.g. with the mouse), the
window states would be updated repeatedly.

To resolve these issues, the window states are no longer updated when
the LVN_ITEMCHANGED notification is received. Instead, a timer is set
and the window states are updated when the timer elapses. Repeated
LVN_ITEMCHANGED notifications simply cause the timer to be reset.

This has the advantage of being simpler and handling any of the ways in
which a group of items might be selected. The timer will fire after
50ms, which should be short enough to not be noticeable.