Fix code that prevents listview selection on right click
In the "list" view mode, clicking a row (either via left-click or
right-click) will select the entire row by default. For right-click,
this is undesirable, as the row should only be selected if the filename
is right-clicked. Right-clicking a row outside of the filename should
result in the background context menu being shown.

The LVN_ITEMCHANGING handler wasn't always correctly blocking the
selection. For example, if you selected one row then right-clicked an
empty spot on another row, the second row would be selected.

Therefore, the LVN_ITEMCHANGING handler has been updated so that it
specifically blocks an item from transitioning from not selected to
selected when necessary. This should mean that right-clicking an empty
spot on a row results in the background context menu being consistently
shown.