Add initial BookmarkItem class
This class will ultimately be used to represent a bookmark item (either
a bookmark or bookmark folder). It has a few notable differences from
the two existing classes (CBookmark and CBookmarkFolder):

- Only a single class is used to represent bookmarks and bookmark
  folders. This is ultimately simpler and it means that child items
  don't have to be stored using a variant type.
- The class is marked as non-copyable. Each bookmark item has a unique
  ID. No other bookmark item shares the same ID. Therefore, copying a
  bookmark item isn't something that makes sense. That would allow two
  copies of a bookmark object to exist, with the same ID, but different
  properties.
- Child items are stored using unique_ptr. This should make it easier to
  construct and insert items. In the current system, bookmark objects
  can be created directly or via a dynamically allocated object. In the
  updated system, all bookmark items will be set up as unique_ptrs.

Note that the BookmarkItem class isn't currently used anywhere and
hasn't been tested at all yet. Because of how the current bookmark
classes are used, it should be easier to create a new class and switch
over to that, rather than trying to merge the two existing classes.