Show bookmarks in main menu
To make things simpler, the entire bookmarks menu is now dynamically
generated. At the moment, the entire menu is regenerated every time the
main menu is opened, but it's possible that may be optimised in future
(e.g. there's no need to regenerate the menu if no bookmark items have
changed since the menu was last generated).