Update way in which manage bookmark dialog navigates
The dialog no longer handles listview double-clicks or treeview
selection change notifications. Instead, these are handled by the
listview and treeview, respectively.

When the listview navigates (in response to a double-click), the dialog
will update the treeview selection as necessary. Correspondingly, when
the treeview selection changes, the dialog will request that the
listview navigate.