Update functions used to read/write data from/to HGLOBAL
There are currently two forms of data that can be written to a HGLOBAL:

- A string
- Binary data (stored within a string)

Because both of these are stored using strings, there was previously
only a single function used to write either of these to a HGLOBAL.
However, a string needs to have the null terminator written, while the
binary data doesn't. Because there was only a single function being
used, it was always writing a null terminator, even when it wasn't
necessary.

To fix this, there are now two functions for writing to a HGLOBAL: one
for strings and one for binary data. Additionally, this should make the
code clearer overall because it means that it's obvious when you're
writing binary data (as opposed to it being implied by the situation).