Indicate dropping a bookmark is supported on bookmarks toolbar
When dragging a bookmark item over the bookmarks toolbar, the drop
effect will now be set to indicate that the item can be moved
(previously, the drop effect was set to none, which indicated that a
bookmark item couldn't be dropped).

Note that it's not yet possible to actually move a bookmark item by
dropping it. This commit purely affects the drop feedback that's
provided as a bookmark item is dragged over the toolbar.