Store original GUIDs when serializing bookmark items
This will be needed in order to allow a bookmark item to be dragged and
dropped within the application. When a drag of that type occurs, the
original bookmark will be serialized and stored within the IDataObject
associated with the drag. Once the bookmark is dropped, it's the
original bookmark that needs to be moved, not the copy stored within the
IDataObject instance.

Therefore, when serializing a bookmark item, the original GUID is now
stored and made available on the deserialized bookmark item. This GUID
can then be used to locate the original bookmark item.

The original GUID will be cleared when a bookmark item is added to the
bookmark tree (as at that point, the information should no longer be
necessary).