Adjust drop behaviour for folders on bookmarks toolbar
When an item is dragged over the bookmarks toolbar, the position of the
item (if it were to be dropped) is shown using an insertion mark. If the
dragged item is over a bookmark, then the position is before the
bookmark if the item is over the first half of the item, and after the
bookmark if the item is over the second half of the item.

Previously, if an item was dragged over a bookmark folder, the drop
location would always be marked as the folder itself, no matter the
location of the dragged item.

Because there's not currently any spacing between buttons, this was an
issue. It meant that you couldn't directly position a dragged item
before/after a folder.

To resolve this, the drop position is now set as follows:

- To the position before the folder if the dragged item is within the
  first 20% of the button.
- To the folder itself if the dragged item is within the middle 60% of
  the button.
- To the position after the folder if the dragged item is within the
  last 20% of the button.