Only handle DragOver() in bookmarks treeview when necessary
The insertion mark shown in the bookmarks treeview while dragging may
flicker as the mouse moves. To help prevent this (and to save some
unnecessary work), the DragOver() now returns immediately if at least
one of the following conditions is true:

- The mouse position hasn't changed. In this case, the drop target is
  guaranteed to be the same as before.
- The drop target hasn't changed. In this case, the mouse may have
  moved, but the final target is still the same as it was previously.

In either of the above cases, there's no need to update the insertion
mark.