Simplify bookmark target data that's stored during drag
Storing a reference to the selected item is redundant in the bookmarks
listview and bookmarks toolbar (since it can be determined if
necessary).

However, the main point of this change is to make it easier to abstract
some of the bookmark drop code out. Ensuring that the windows that
accept a bookmark drop don't need to store any additional information
about the drop target simplifies the resulting implementation.