Show context menu when right-clicking bookmarks listview background
With this change, the add bookmarks dialog will now also select the
appropriate default parent folder when necessary (e.g. if you add a new
bookmark using the context menu in a specific folder).

Additionally, if you right-click a bookmark folder, it will now be
selected as the parent folder for any context menu operations.
Previously, the parent folder of the currently selected item(s) would
always be used.