Show context menu when right-clicking bookmark folder in menu
Previously, when the bookmarks menu was being shown, the context menu
would be shown when right-clicking a bookmark. However, right-clicking a
folder would have no effect.

The reason for this behaviour was that menu items that open submenus
don't have IDs. This meant that it wasn't possible to look the bookmark
item up, so it wasn't possible to determine which item had been clicked.

To resolve that, a set of menu position mappings are now stored when
building the bookmarks menu. These map between a menu + position pair
and a bookmark item. This means that it's now possible to determine
which item was right-clicked, regardless of whether it was a bookmark or
folder.

Note that, from some testing, it appears that the WM_MENURBUTTONUP
message isn't sent when right-clicking an item whose submenu is being
shown. The message will be sent if the item is right-clicked when the
submenu isn't being shown.