Add ability to bookmark all tabs
This function can be invoked via a "Bookmark All Tabs..." item that has
been added to the main bookmarks menu.

Note that this menu item has been assigned a keyboard shortcut of
Ctrl+Shift+D. This shortcut is consistent with Chrome and is similar to
the shortcut used to bookmark a single tab (Ctrl+D).

Note, however, this keyboard shortcut was already being used by the
wildcard deselect menu item. Therefore, the shortcut combinations for
the two wildcard menu items have been changed as follows:

- Wildcard select: Ctrl+Alt+S
- Wildcard deselect: Ctrl+Alt+D