Show folder icons for bookmarks on bookmarks toolbar
Previously, the bookmarks toolbar always used a single icon for all
bookmarks. Now, the icon for each bookmark is retrieved in the
background and the toolbar is updated when this process completes.

The default icon for each bookmark has also been changed to the default
system folder icon (since this is the icon most folders in the system
will likely have). This does mean that bookmark folders and bookmarks
now share similar icons (they're both folders), which may be worth
looking at in the future.

This ties into a more general strategy, though, to always show icons for
folders shown in the application.