Fix way in which bookmark locations are copied to clipboard
When copying a set of bookmark/bookmark folders to the clipboard, the
bookmark locations and folder names are written as text (in addition to
the bookmark items themselves being copied), with the text for each item
written on a new line.

Previously, this was being done incorrectly. Each item was being written
as text, so that the text for the next item would overwrite the text for
the previous item.

To fix this, the lines for each item are now joined together and written
once.