Add tests for BookmarkTree class
Note that Explorer++.rc is now included in the TestExplorer++ build.
This is because the BookmarkTree constructor needs to load various
string resources (the names of the permanent bookmark folders).

At the moment, Explorer++_NoTranslation.rc isn't included, as the images
it includes aren't needed.

While the project resources are used in the UI and shouldn't strictly be
needed for unit tests, there's not necessarily an easy way of getting
around that in this case, since you would need an alternative way of
setting up the bookmark tree.

Additionally, including the resources is simple and they may be needed
for other unit tests in the future.