Use std::optional instead of boost::optional
This has the advantage of being consistent. It also means that it's only
necessary to include one header (for std::optional), rather than two.

Note that the only place that std::optional isn't used is in the
Plugins::TabsApi class. sol is designed to work with its own
sol::optional class, so that's what's used there instead (where
necessary).