Update back/forward navigation in manage bookmarks dialog
The dialog now uses NavigationController, in the same way as it's used
in the main part of the application (the common parts of its
functionality have been moved out into a separate class).

This fixes at least one issue with the back/forward system: previously,
the back button would be incorrectly enabled after the first navigation.
It also means that a list of back/forward items can be shown in future
if desired.