Add .clang-format file
This file has been setup to produce results that are close to the
current code style used within the project. Note that the results aren't
exact (e.g. because clang-format doesn't always allow the style of a
particular piece of code to be configured), but should be broadly
similar.

Aside from being able to easily format code on demand, having this file
is also useful in that it explicitly lists the recommended formatting
and style options.