Fix issue when updating headers via listview header context menu
In ef4f8763cbeee7329e81f6ee2789b65c4363cdf8, the listview header context
menu code was moved to the ShellBrowser class. As part of that change,
the code to handle a menu item selection was also moved.

The code that handles the selection, however, wasn't updated to handle
selection of the "More..." item; it was simply assumed that any menu
item would correspond to a column. This then meant that if you selected
the "More..." item, the application would crash. This was reported in
https://explorerplusplus.com/forum/viewtopic.php?f=5&t=2591.

Therefore, the item selection code has been updated to handle the
"More..." item. As part of fixing this issue, another issue has also
been fixed in which the sort by/group by menus wouldn't be updated when
adding or removing a column via the header context menu.