Allow version text in about dialog to be selected
Previously, the version information text (version number + build date)
was contained within two static controls. This meant that the text
couldn't be selected or copied.

Now, the text is contained within a read only edit control, which does
allow the text to be selected.

Note that, as part of this change, the "Unicode" text has been removed
from the version information, since it's redundant (all builds are
Unicode builds).

This closes #188.