Fix issue with overlapping private window messages
Previously, WM_USER_NEWITEMINSERTED overlapped with
WM_APP_ICON_RESULT_READY. This could then cause issues when one of the
messages was sent.

To fix this and to help avoid future clashes, WM_APP_ICON_RESULT_READY
has been redefined so that it's at the upper end of the WM_APP range.