Add specific configuration for beta and release branches in AppVeyor configuration
A branch whose name ends in "-beta" is considered to be a branch for a
beta release, while a branch whose name ends in "-release" is considered
to be a branch for a stable release.

When one of these branch names is used, an appropriate environment
variable will be set, which will result in a corresponding preprocessor
define being set within the Explorer++ project. That way, it's possible
to determine at compile time the type of release that's being built.

This is in preparation for being able to build and manage release
versions on AppVeyor.