Update way in which COM is initialized in background threads
Previously, COM was initialized by pushing a task to the relevant worker
threads. However, when a navigation occurs, the queue of tasks for each
of the worker threads is cleared. This includes the first navigation for
a tab. Because of that, the COM initialization task could be cleared
from the queues before it ever had the chance to run.

Depending on the timing of the items above, this could cause icon
retrieval to fail. It could also the application to crash in strange
ways, usually within ole32.dll.

To fix this, the CTPL library has been customised to provide callbacks
for when a thread is starting/finishing. Modifying a third-party library
isn't ideal, but it was done for two reasons in this case:

- The implementation is likely simpler than the alternatives (e.g.
  trying to keep track of whether CoInitializeEx has been called, or
  just calling CoInitializeEx whenever necessary and never calling
  CoUninitialize).

- CTPL hasn't received any commits in about five years, so it's unlikely
  it's going to change much (or at all) in the future. So it's probably
  not going to be necessary to reapply these changes later on.