Set force_update to true in deploy config
This is necessary, as explorer++_translations.zip is created twice,
added as an AppVeyor artifact twice, and deployed twice. Attempting to
deploy the same file more than once will result in an error, though, so
by setting force_update to true, the file will simply be overwritten the
second time it's deployed.