Move deployment section out of branch-specific area
When building from a tag, the associated branch may not be clear (e.g.
the tag might exist on several different branches). So deploying only
when a tag is pushed on a specific branch isn't necessarily going to
work.

To resolve this, the branch-specific section has been removed and the
deployment section has been moved into the general section (though
deployment will still only be triggered on a tag push).