Show warning message when setting Explorer++ as the default file manager
As there is no need to install Explorer++, it's possible to place it in
any location, then set it as the default file manager. Removing the
executable without first reverting the default file manager setting is
problematic, as the associated registry key will still point to the
original location of the Explorer++ executable. As there's no installer,
removing the executable is trivial and it's not necessarily clear that
you need to revert the setting first.

To try and prevent potential issues, a warning is now shown when
attempting to set Explorer++ as the default file manager. This warning
requires the user to explicitly proceed. Additionally, the default
button in the warning dialog is "No", which will cancel the process.