Update Sol to latest (head) version
This is to ensure that the code compiles in Visual Studio 16.6.0. In the
latest released version of Sol, there's an error triggered while
compiling (see https://github.com/ThePhD/sol2/issues/935).

Updating to a non-released version isn't ideal, but doesn't matter too
much, as the functionality Sol is used to provide is still in
development.