Fix command line parameter issue
When Explorer++ is set as the default file manager, it can be invoked in
the following way:

C:\path\to\Explorer++.exe "C:\"

Due to how command line arguments are parsed, the path argument above
will be interpreted as:

C:"

This will then prevent the directory from being opened.

To resolve this issue, any directory path that ends in a double quote
character will now have that character replaced with a backslash. This
should be safe, as the quote character isn't a valid filename character,
so should never legitimately appear in a directory path.

This fixes #202. This issue was also reported at
https://explorerplusplus.com/forum/viewtopic.php?f=5&t=2406.

(cherry picked from commit dc32d016c72ade0363432f9826d0520ac99397e0)