Clamp values passed via WM_USER_DISPLAYWINDOWRESIZED
When the display window is resized, a WM_USER_DISPLAYWINDOWRESIZED
message is sent to the parent. This message contains the updated
width/height of the display window.

With the way the updated width and height are calculated, it's possible
for the values to be negative. This was causing problems because the
values would be interpreted as unsigned on the receiving end and so a
negative value would be reinterpreted as a different positive value.

To fix this, the values are now clamped before they're sent so that
they're in the positive range. That is, the minimum value that will be
sent is 0.

This fixes #233.

(cherry picked from commit 77107493ee9ec2fac2f2d46b04783e88f4cfe0f7)