Correctly handle case in which name column isn't first
When the listview is in details mode, the position of the name column
can be changed. By default, it's the first column, but it can be moved
to another position.

Previously, if the column was moved to another position, the first
column would still always have its text so to the name of each item.
This was because the text in the first column was always fixed (to the
associated item's name).

This was essentially a simplification in the code, as the text in the
first column (which corresponds to an iSubItem of 0) is the item label
used in every other view.

Now, this situation is correctly handled. The text in the first column
will be appropriately set when inserting items. It will also be reset
(when necessary) when switching between details view/non-details view.

This fixes #194.

(cherry picked from commit 309d64591805d4c3fd545c4f55eed01846487dd6)