Add initial dark mode code
This commit adds some initial support for theming the application in
dark mode.

Dark mode theming relies on two basic things:

1. The use of undocumented APIs to change the title bar and menu colors.
2. Custom code within the application to change background and text
   colors for each of the windows.

Because of the use of undocumented APIs, this functionality is currently
behind a command line flag (--enable-dark-mode).

At the moment, the following elements are styled:

- The title bar.
- All menus shown within the application.
- The treeview.
- Each listview.

All other elements (e.g. the toolbars, the dialogs, the status bar) will
still have their original colors.

This is part of the initial work for issue #115.