Set rebar background color in dark mode
As each of the toolbars has the TBSTYLE_TRANSPARENT style set, the rebar
background color will appear underneath each toolbar.

Note that, at the moment, the toolbar text colors aren't changed, so the
text in each toolbar will still be the original color.

This is part of the work for issue #115.