Update dialogs for dark mode
With this commit, each dialog will now be styled for dark mode when
appropriate, with most of the work being done in a common subclass. Note
that not all controls are styled correctly. However, the dialogs mostly
look like what you'd expect in dark mode.

Also note that the options dialog hasn't been updated and will still
display in the original colors.

This is part of the work for issue #115.