Set dark mode theme for status bar
The theme is set to "ExplorerStatusBar" (which will become
"DarkMode::ExplorerStatusBar" when dark mode is on). Ultimately, it
would probably be more reliable to customise the status bar manually in
dark more, but that would be more work, while the solution here is
trivial.

The only thing that doesn't appear to render correctly with this theme
is the size grip.

This is part of the work for issue #115.