Update address bar dark mode background color
As with the treeview, the updated color is the same as that used in
Windows Explorer.

Note that the theme that is now used with the address bar combobox
("AddressComposited") has a raised dropdown button. Because that button
is raised, it doesn't have a border around it. That's a relatively small
issue, however, and this theme gives results that are closer to those
seen in Explorer.

Within Windows, this theme appears to be used when the user clicks in
the address bar. The address bar switches from a toolbar to a combobox
that's styled in a similar way.

This is part of the work for issue #115.